+++ RayTracer/Engine/Material.hpp
@@ -6,6 +6,6 @@ struct Material
 {
 public:
 	Color	color = Color(128, 128, 128);
-	float	ambient = 0.1f;
-	float	reflectivity = 0.2f;
+	double	ambient = 0.1;
+	double	reflectivity = 0.2;
 };
